marca

+++ locadora_carros/app/Http/Controllers/MarcaController.php
@@ -9,6 +9,11 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+   
+    public function __construct(Marca $marca) {
+        $this->marca = $marca;
+    }
+   
     /**
      * Display a listing of the resource.
      *
@@ -16,19 +21,10 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
-        return $marcas;
+        $marcas = $this->marca->all();
+        return response()->json($marcas, 200);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -39,54 +35,97 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
 
-        $marca = Marca::create($request->all());
-        return $marca;
+        $this->validate($this->marca->rules(), $this->marca->feedback());
+
+        $marca = $this->marca->create($request->all());
+        return response()->json($marca, 200);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-       return $marca;
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Marca  $marca
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Marca $marca)
-    {
-        //
+        $marca = $this->marca->find($id);
+        
+        if($marca === null){
+            
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404);
+       
+        }
+        
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+        
+        if($marca === null){
+            
+            return reponse()->json(['erro' => 'Impossível realizar atualização'], 404);
+       
+        }
+
+        if($request->method() === 'PATCH') {
+
+            $regrasDinamicas = array();
+
+           // $teste = '';
+            
+            foreach($marca->rules() as $input => $regra) {
+                
+               // $teste .= 'Input'.$input. '| Regra:'.$regra.'<br>';
+                
+                if(array_key_exists($input, $request->all())) {
+                    
+                    $regrasDinamicas[$input] = $regra;
+                
+                }
+                
+            }
+            
+            $request->validate($regrasDinamicas, $marca->feedback());
+
+        }else {
+           
+            $request->validate($marca->rules, $marca->feedback());
+       
+        }
+
+        
         $marca->update($request->all());
-        return $marca;
+        response()->json($marca, 200);
+    
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->fid($id);
+        
+        if($marca === null){
+            
+            return response()->json(['erro' => 'Impossível realizar o delete'], 404);
+        
+        }
+
         $marca->delete();
-        return ['msg' => 'Deletado'];
+        return response()->json(['msg' => 'Deletado'], 200);
     }
 }
